% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFtoFF.R
\name{DFtoFF}
\alias{DFtoFF}
\title{Data frame to flowFrame}
\usage{
DFtoFF(DF)
}
\arguments{
\item{DF}{A data frame or list of data frames to be converted}
}
\value{
A flowFrame or list of flowFrames
}
\description{
Utility function for coercing data frames to the 'flowFrame' S4 class, which many flow
  packages are designed to work with.
}
