% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFtoDF.R
\name{FFtoDF}
\alias{FFtoDF}
\title{Convert a flowFrame to a data frame}
\usage{
FFtoDF(FF)
}
\arguments{
\item{FF}{A flowFrame or list of flowFrames to be converted}
}
\value{
A data frame or list of data frames
}
\description{
Utility function for coercing S4-type flowFrame objects to data frames
}
