% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_CompPlot.R
\name{flow_CompPlot}
\alias{flow_CompPlot}
\title{Visually Inspect Compensation Samples}
\usage{
flow_CompPlot(compList, parameter, compMatrix)
}
\arguments{
\item{compList}{A list of flowFrames containing compensation control data. Samples may need
pre-processing (to remove doublets, outliers) to generate nice-looking plots.}

\item{parameter}{Optional character string defining a single compensation sample (e.g., PE control) to build a plot for. 
If not provided, defaults to building plots for all items in compList.}

\item{compMatrix}{Optional data frame containing compensation values. By default, the compensation matrix which is applied to
compList should be the matrix that was built during acquisition to compensate the actual experimental samples.
This can be easily obtained by calling the get_CompMatrix() function on a flowFrame containing experimental data, but
only if the flowFrame has never been converted to a data frame, as this drops all experimental metadata associated with
the flowFrame class.}
}
\description{
A function for building flowJo V10-like plots that allow for visual inspection of
the quality of acquired compensation matrices
}
