% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_MarkerApply.R
\name{flow_MarkerApply}
\alias{flow_MarkerApply}
\title{Apply informative marker labels to imported flow data}
\usage{
flow_MarkerApply(flowObj, keepFluoro = FALSE, sepChar = "-")
}
\arguments{
\item{keepFluoro}{(Optional) Logical, if TRUE appends the marker label to the fluorescence label}

\item{sepChar}{(Optional, dependent on keepFluoro = TRUE) String defining the character separating
the marker and fluorescene labels. Default = "-"}
}
\description{
Using the parameter table called by the get_Parameters() function, applies
  marker labels to flow data frames.
}
