% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_OptimalClust.R
\name{flow_OptimalClust}
\alias{flow_OptimalClust}
\title{Optimal Number of Clusters for Flow Data}
\usage{
flow_OptimalClust(flowObj, algorithm, kMax, nBoot, plot = TRUE)
}
\arguments{
\item{flowObj}{A data frame or flowFrame to be analyzed}

\item{algorithm}{The clustering algorithm to utilize. Options (no quotations): kmeans, pam (Partitioning Around Mediods), clara (Clustering LARge Applications), hcut (Hierarchical clustering, cut tree into 'k' clusters)}

\item{kMax}{Integer representing max number of possible clusters to consider}

\item{nBoot}{Integer representing number of times to repeat the analysis from a random starting position, must be >1 for gap statistic analysis}

\item{plot}{Logical indicating whether or not to generate graphical plots for wss, silhouette, and gap statistic methods}
}
\description{
Function implementing various statistical methods to determine the optimal number of clusters for flow cytometry data,
  using various clustering algorithms. ***WARNING*** extremely computationally intensive-- very large data sets may not
  function and/or may take significant time to complete.
}
