% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_Prune.R
\name{flow_Prune}
\alias{flow_Prune}
\title{Prune Flow Data}
\usage{
flow_Prune(flowObj, parameter, upperBound = NULL, lowerBound = NULL,
  fromTable = FALSE)
}
\arguments{
\item{flowObj}{Data frame, flowFrame, or a list containing either class of object to be manipulated}

\item{parameter}{String or vector of strings matching a column label or set of labels in the flow data frame}

\item{upperBound}{(Optional) Numeric defining the allowed maximum value for a specific parameter (if designated by the parameter argument)
or for any parameter (if no parameter argument provided). If a value falls outside this bound, the entire event
  (row) will be removed from the flow data frame.}

\item{lowerBound}{See upperBound}

\item{fromTable}{(Optional, dependent on calling get_Parameters() function) Logical indicating whether
or not the function should look in the parameter table for upper/lower bound arguments. If true,
ignores provided parameter, upperBound, and lowerBound arguments (all the logic is in the table).
Default = FALSE}
}
\value{
A pruned flow object of the type initially passed to the function
}
\description{
Function to remove columns or events that meet specific conditions from a flow data frame.
}
