% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_Subset.R
\name{flow_Subset}
\alias{flow_Subset}
\title{Subset Flow Data}
\usage{
flow_Subset(flowObj, subsetPercent)
}
\arguments{
\item{flowObj}{A data frame, flowFrame, or a list containing these data types}

\item{subsetPercent}{Numeric from 0-1 denoting the percentage of the original data set to return}
}
\value{
A subsetted flow object of the type initially passed to the function
}
\description{
Simple function which returns a random subset of values from the flow object it is passed.
}
