% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CompMatrix.R
\name{get_CompMatrix}
\alias{get_CompMatrix}
\title{Retrive Compensation Matrix}
\usage{
get_CompMatrix(flowFrame)
}
\arguments{
\item{flowFrame}{An object of the flowFrame class which possess compensation metadata.}
}
\value{
A data frame of compensation values.
}
\description{
Reads the metadata of a flowFrame object and returns the compensation matrix which was applied to it.
The structure of the returned data frame: Column labels denote fluorescence channel, while
row labels denote spillover from other fluorophores detected in that channel.
The returned compensation matrix can be edited and reapplied to any flowFrame using the flowCore function
compensate(flowFrame, compensationMatrix), as long as the fluorophore labels match.
}
