% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FCS.R
\name{get_FCS}
\alias{get_FCS}
\title{FCS File Importer}
\usage{
get_FCS(pattern, path, asDF = FALSE)
}
\arguments{
\item{pattern}{(Optional) Character string designating a file name pattern.
Only files that match the pattern will be imported.}

\item{path}{(Optional) Character string designating the file path in which to
look for FCS files}

\item{asDF}{Logical value representing whether or not to convert the incoming
FCS files to dataframes, instead of flow frames. Defaults to FALSE.}
}
\value{
A list of data frames or flow frames.
}
\description{
Generic function for pulling in FCS files and returning a list of flow frames.
  Default behavior is to import all .fcs files in the working directory. Names each element of the list using the associated file name, minus the ".fcs" file extension.
}
