% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ParameterTable.R
\name{get_Parameters}
\alias{get_Parameters}
\title{Parameter Table Importer}
\usage{
get_Parameters(parameterTableFileName)
}
\arguments{
\item{parameterTableFileName}{Character string representing file name of parameter table in working directory.}
}
\description{
Function for reading in a tab-delimited file from the working directory containing
  labels for fluorophores and markers used in a flow experiment. The table is used
  by the flow_MarkerApply() function to apply more descriptive and informative labels
  to flow data frames. Optionally allows for description of upper and lower parameter
  bounds for removing unwanted data from flow frames using the flow_Prune() function.
  Table should have a header with the following labels:
  Column 1) "Fluorophore"
  Column 2) "Marker"
  Optional Column 3) "UpperBound"
  Optional Column 4) "LowerBound"
  Fill empty cells (e.g., marker value for FSC, or no argument for upper/lower bound) with "NA".
  The parameter labels themselves require no specific order/punctuation/casing for
  fluorophores or markers, but the fluorophore names should match those applied to
  the FCS files being worked with. Entering an Upper and LowerBound value of "0" for
  a parameter indicates that that parameter should be dropped when the flow_Prune() function
  is called.
}
